feat(codegen): move under internal/cmd
It should not be exported for anyone to use. It's meant purely
for the sumup-go SDK.
