+++ .github/workflows/codegen-ci.yml
@@ -6,13 +6,13 @@ on:
     branches:
       - main
     paths:
-      - 'codegen/**'
+      - 'internal/cmd/codegen/**'
       - '.github/workflows/codegen-ci.yml'
   pull_request:
     branches:
       - main
     paths:
-      - 'codegen/**'
+      - 'internal/cmd/codegen/**'
       - '.github/workflows/codegen-ci.yml'
 
 env:
@@ -27,13 +27,13 @@ jobs:
     runs-on: ubuntu-latest
     defaults:
       run:
-        working-directory: codegen
+        working-directory: internal/cmd/codegen
     steps:
       - uses: actions/checkout@8e8c483db84b4bee98b60c0593521ed34d9990e8 # v6.0.1
 
       - uses: actions/setup-go@4dc6199c7b1a012772edbd06daecab0f50c9053c # v6.1.0
         with:
-          go-version-file: codegen/go.mod
+          go-version-file: internal/cmd/codegen/go.mod
 
       - name: golangci-lint
         uses: golangci/golangci-lint-action@1e7e51e771db61008b38414a730f564565cf7c20 # v9.2.0
@@ -45,13 +45,13 @@ jobs:
     runs-on: ubuntu-latest
     defaults:
       run:
-        working-directory: codegen
+        working-directory: internal/cmd/codegen
     steps:
       - uses: actions/checkout@8e8c483db84b4bee98b60c0593521ed34d9990e8 # v6.0.1
 
       - uses: actions/setup-go@4dc6199c7b1a012772edbd06daecab0f50c9053c # v6.1.0
         with:
-          go-version-file: codegen/go.mod
+          go-version-file: internal/cmd/codegen/go.mod
 
       - name: Test
         run: go test -v -failfast -race -timeout 10m ./...
