+++ internal/cmd/codegen/internal/strcase/strings.go
