+++ internal/cmd/codegen/pkg/builder/http.go
