+++ internal/cmd/codegen/pkg/builder/intermediate_representation.go
