+++ internal/cmd/codegen/pkg/builder/resolve.go
