+++ internal/cmd/codegen/pkg/builder/transform.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/getkin/kin-openapi/openapi3"
 
-	"github.com/sumup/go-sdk-gen/internal/strcase"
+	"github.com/sumup/sumup-go/internal/cmd/codegen/internal/strcase"
 )
 
 // schemasToTypes converts openapi3 schemas to golang struct and enum types.
