+++ internal/cmd/codegen/pkg/builder/types.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/getkin/kin-openapi/openapi3"
 
-	"github.com/sumup/go-sdk-gen/internal/strcase"
+	"github.com/sumup/sumup-go/internal/cmd/codegen/internal/strcase"
 )
 
 type Writable interface {
