+++ internal/cmd/codegen/templates/base.go.tmpl
