+++ internal/cmd/codegen/templates/client.go.tmpl
