+++ internal/cmd/codegen/templates/resource.go.tmpl
